requirements.txt
# Core Data Science Libraries
numpy==1.24.3
pandas==2.0.3
scipy==1.11.1

# Backtesting & Trading
backtrader==1.9.78.123

# Data Sources
yfinance==0.2.28

# Visualization
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.15.0

# Configuration
pyyaml==6.0.1
python-dotenv==1.0.0

# Technical Analysis
pandas-ta==0.3.14b

# Utilities
pytz==2023.3
tqdm==4.65.0

# Testing
pytest==7.4.0

# Logging
loguru==0.7.0
